requirements.txt
python3
pip
discord.py==2.0.1
python-dotenv==0.21.0
requests==2.25.1
PyNaCl==1.5.0